移动目录，并添加注册，登陆页面
