--- src/renderer/components/Header/WaHeader.vue
@@ -1,30 +1,0 @@
-<template>
-  <el-menu :default-active="activeIndex" type="flex" justify="end" class="el-menu-demo" mode="horizontal" @select="handleSelect">
-    <el-menu-item index="1">处理中心</el-menu-item>
-    <el-submenu index="2">
-      <template slot="title">我的工作台</template>
-      <el-menu-item index="2-1">选项1</el-menu-item>
-      <el-menu-item index="2-2">选项2</el-menu-item>
-      <el-menu-item index="2-3">选项3</el-menu-item>
-      <el-submenu index="2-4">
-        <template slot="title">选项4</template>
-        <el-menu-item index="2-4-1">选项1</el-menu-item>
-        <el-menu-item index="2-4-2">选项2</el-menu-item>
-        <el-menu-item index="2-4-3">选项3</el-menu-item>
-      </el-submenu>
-    </el-submenu>
-    <el-menu-item index="3" disabled>消息中心</el-menu-item>
-    <el-menu-item index="4"><a href="#" target="_blank">订单管理</a></el-menu-item>
-  </el-menu>
-</template>
-<script>
-export default {
-  name: 'wa-header'
-}
-</script>
-<style type="text/css">
-/* .el-menu-demo{
-  float: right;
-  
-} */
-</style>
