+++ src/renderer/components/Panel/Main/Body/RawData.vue
